staging/src/k8s.io/apimachinery/pkg/util/net/util.go
/*
Copyright 2016 The Kubernetes Authors.

Licensed under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

    http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.
*/

package net

import (
	"context"
	"errors"
	"net"
	"reflect"
	"syscall"
	"time"
)

// IPNetEqual checks if the two input IPNets are representing the same subnet.
// For example,
//	10.0.0.1/24 and 10.0.0.0/24 are the same subnet.
//	10.0.0.1/24 and 10.0.0.0/25 are not the same subnet.
func IPNetEqual(ipnet1, ipnet2 *net.IPNet) bool {
	if ipnet1 == nil || ipnet2 == nil {
		return false
	}
	if reflect.DeepEqual(ipnet1.Mask, ipnet2.Mask) && ipnet1.Contains(ipnet2.IP) && ipnet2.Contains(ipnet1.IP) {
		return true
	}
	return false
}

// Returns if the given err is "connection reset by peer" error.
func IsConnectionReset(err error) bool {
	var errno syscall.Errno
	if errors.As(err, &errno) {
		return errno == syscall.ECONNRESET
	}
	return false
}

// Returns if the given err is "connection refused" error
func IsConnectionRefused(err error) bool {
	var errno syscall.Errno
	if errors.As(err, &errno) {
		return errno == syscall.ECONNREFUSED
	}
	return false
}

// ExceedDeadlineOnCancel cancels the connection upon context cancellation. It does so by setting an exceeded deadline
// on the connection, which leads to a timeout. If the connection should not have a deadline in the end, the returned
// function will remove it again. That function is usually deferred.
func ExceedDeadlineOnCancel(ctx context.Context, conn net.Conn) (removeDeadline func()) {
	done := make(chan struct{})

	go func() {
		<-ctx.Done()
		_ = conn.SetDeadline(time.Unix(0, 0))
		close(done)
	}()

	return func() {
		<-done
		_ = conn.SetDeadline(time.Time{})
	}
}
